Use libram requirement, add nobarf option (#455)
* Remove lib dependency from combat

* Add nobarf option, move estimatedTurns calc

* Address nobarf review comments
